Initial Commit
This is my first try of using git in kotlin learning.
C:\Users\Owner\IdeaProjects\kotlin-starter>git add *
warning: in the working copy of '.gitignore', LF will be replaced by CRLF the next time Git touches it
warning: in the working copy of 'gradle.properties', LF will be replaced by CRLF the next time Git touches it
warning: in the working copy of 'gradlew', LF will be replaced by CRLF the next time Git touches it
warning: in the working copy of 'src/main/kotlin/Main.kt', LF will be replaced by CRLF the next time Git touches it# with '#' will be ignored, and an empty message aborts the commit.